Convert getRandomIndex and getSessionsId utils to TypeScript
