--- src/utils/getRandomIndex.js
@@ -1,7 +1,0 @@
-export default function getRandomIndex(array) {
-    if(array.length === 0) {
-      return null; // Return null if the array is empty
-    }
-    const randomIndex = Math.floor(Math.random() * array.length);
-    return array[randomIndex];
-  }
